#3612: extract shared DailyInvoiceImportJobBase to dedupe CZK/EUR import jobs
DailyInvoiceImportCzkJob and DailyInvoiceImportEurJob were byte-identical
apart from currency, cron, and metadata strings, so a fix to one (e.g. the
date calculation or error handling) risked being forgotten in the other.
Extracts the shared ExecuteAsync/ImportYesterdayForCurrency logic into an
abstract DailyInvoiceImportJobBase, following the existing
BankImportJobBase/ComgateCzkImportJob pattern used elsewhere in the codebase.

Closes #3612
